+++ R/multipliers.R
@@ -9,7 +9,7 @@
 #' \code{class} 'uecm'. This is because of the different (but equivalent)
 #' transformation functions that are used for each class/model ('ardl' and
 #' 'uecm') to estimate the multipliers.
-#' 
+#'
 #' Note that \code{type = 0} is equivalent to \code{type = "sr"}. Also,
 #' \code{type = s} will produce the same estimates as \code{type = "lr"} for
 #' those particular variable for which s >= from their ARDL order.
@@ -40,53 +40,53 @@
 #' }
 #' \deqn{\mu = \frac{c_{0}}{1-\sum_{i=1}^{p}b_{y,i}}}
 #' \deqn{\delta = \frac{c_{1}}{1-\sum_{i=1}^{p}b_{y,i}}}
-#' 
+#'
 #' \describe{
 #'   \item{As derived from an Unrestricted ECM:}{}
 #' }
 #' \deqn{\mu = \frac{c_{0}}{-\pi_{y}}}
 #' \deqn{\delta = \frac{c_{1}}{-\pi_{y}}}
-#' 
+#'
 #' \strong{Short-Run Multipliers:}
 #' \describe{
 #'   \item{As derived from an ARDL:}{}
 #' }
 #' \deqn{\frac{\partial y_{t}}{\partial x_{j,t}} = \frac{b_{j,0}}{1-\sum_{i=1}^{p}b_{y,i}} \;\;\;\;\; \forall j=1,\dots,k}
-#' 
+#'
 #' \describe{
 #'   \item{As derived from an Unrestricted ECM:}{}
 #' }
 #' \deqn{\frac{\partial y_{t}}{\partial x_{j,t}} = \frac{\omega_{j}}{-\pi_{y}} \;\;\;\;\; \forall j=1,\dots,k}
-#' 
+#'
 #' \strong{Interim Multipliers:}
 #' \describe{
 #'   \item{As derived from an ARDL:}{}
 #' }
 #' \deqn{\frac{\partial y_{t+s}}{\partial x_{j,t}} = \frac{\sum_{l=1}^{s}b_{j,l}}{1-\sum_{i=1}^{p}b_{y,i}} \;\;\;\;\; \forall j=1,\dots,k \;\;\;\;\; s \in \{0,\dots,q_{j}\}}
-#' 
+#'
 #' \describe{
 #'   \item{As derived from an Unrestricted ECM:}{}
 #' }
 #' \deqn{\frac{\partial y_{t+s}}{\partial x_{j,t}} = \frac{\pi_{j} + \psi_{j,s}}{-\pi_{y}} \;\;\;\;\; \forall j=1,\dots,k \;\;\;\;\; s \in \{1,\dots,q_{j}-1\}}
-#' 
+#'
 #' \strong{Long-Run Multipliers:}
 #' \describe{
 #'   \item{As derived from an ARDL:}{}
 #' }
 #' \deqn{\frac{\partial y_{t+\infty}}{\partial x_{j,t}} = \theta_{j} = \frac{\sum_{l=0}^{q_{j}}b_{j,l}}{1-\sum_{i=1}^{p}b_{y,i}} \;\;\;\;\; \forall j=1,\dots,k}
-#' 
+#'
 #' \describe{
 #'   \item{As derived from an Unrestricted ECM:}{}
 #' }
 #' \deqn{\frac{\partial y_{t+\infty}}{\partial x_{j,t}} = \theta_{j} = \frac{\pi_{j}}{-\pi_{y}} \;\;\;\;\; \forall j=1,\dots,k}
-#' 
+#'
 #' @seealso \code{\link{ardl}}, \code{\link{uecm}}
 #' @author Kleanthis Natsiopoulos, \email{klnatsio@@gmail.com}
 #' @keywords math
 #' @export
 #' @examples
 #' data(denmark)
-#' 
+#'
 #' ## Estimate the long-run multipliers of an ARDL(3,1,3,2) model ---------
 #'
 #' # From an ARDL model
@@ -98,25 +98,25 @@
 #' uecm_3132 <- uecm(ardl_3132)
 #' mult_uecm <- multipliers(uecm_3132)
 #' mult_uecm
-#' 
+#'
 #' all.equal(mult_ardl, mult_uecm)
-#' 
-#' 
+#'
+#'
 #' ## Estimate the short-run multipliers of an ARDL(3,1,3,2) model --------
-#' 
+#'
 #' mult_sr <- multipliers(uecm_3132, type = "sr")
 #' mult_0 <- multipliers(uecm_3132, type = 0)
 #' all.equal(mult_sr, mult_0)
-#' 
-#' 
+#'
+#'
 #' ## Estimate the interim multipliers of an ARDL(3,1,3,2) model ----------
-#' 
+#'
 #' # Note that the estimated interim multipliers match the long-run multipliers
 #' # for those variables that their ARDL order equals or exceeds the interim step
 #' mult_lr <- multipliers(uecm_3132, type = "lr")
 #' mult_1 <- multipliers(uecm_3132, type = 1)
 #' mult_2 <- multipliers(uecm_3132, type = 2)
-#' 
+#'
 #' uecm_3132$order
 #' mult_lr
 #' mult_1
@@ -164,9 +164,9 @@ multipliers.ardl <- function(object, type = "lr", vcov_matrix = NULL) {
         for (i in 1:(length(orders_x) -1)) {
             b0 <- c(b0, b0[length(b0)] + orders_x[i] +1)
         }
-        b0 <- c(1:kw, b0+kw)
+        b0 <- if (kw!=0) c(1:kw, b0+kw) else b0
         # create table without the y in levels and fixed
-        x_table <- dplyr::tibble(name = names(objcoef), coeff = objcoef) %>% 
+        x_table <- dplyr::tibble(name = names(objcoef), coeff = objcoef) %>%
             dplyr::slice(-(from:to)) %>% dplyr::slice(b0)
         # create table only with y in levels
         y_table <- dplyr::tibble(objcoef) %>% dplyr::slice(from:to)
@@ -175,18 +175,18 @@ multipliers.ardl <- function(object, type = "lr", vcov_matrix = NULL) {
         interim = type
         b0 <- 1:(1 + ifelse(orders_x[1] >= interim, interim, orders_x[1]))
         for (i in 1:(length(orders_x) -1)) {
-            b0 <- c(b0, (sum(orders_x[1:i]) + length(orders_x[1:i])+1) : ((sum(orders_x[1:i]) + 
-                length(orders_x[1:i])+1) + 
+            b0 <- c(b0, (sum(orders_x[1:i]) + length(orders_x[1:i])+1) : ((sum(orders_x[1:i]) +
+                length(orders_x[1:i])+1) +
                 ifelse(orders_x[i+1] >= interim, interim, orders_x[i+1])))
         }
-        b0 <- c(1:kw, b0+kw)
+        b0 <- if (kw!=0) c(1:kw, b0+kw) else b0
         # create table without the y in levels and fixed
-        x_table <- dplyr::tibble(name = names(objcoef), coeff = objcoef) %>% 
+        x_table <- dplyr::tibble(name = names(objcoef), coeff = objcoef) %>%
             dplyr::slice(-(from:to)) %>% dplyr::slice(b0)
         # create table only with y in levels
         y_table <- dplyr::tibble(objcoef) %>% dplyr::slice(from:to)
         # create groups to sum by group
-        rep_pattern <- ifelse(orders_x[1:object$parsed_formula$kx] >= interim, 
+        rep_pattern <- ifelse(orders_x[1:object$parsed_formula$kx] >= interim,
             interim, orders_x[1:object$parsed_formula$kx])+1
         x_table <- x_table %>%
             dplyr::mutate(group_id = if (kw != 0) {
@@ -251,7 +251,7 @@ multipliers.uecm <- function(object, type = "lr", vcov_matrix = NULL) {
     } else if ((type == "sr") | (type == 0)) {
         # use gsub because sometimes the spacing goes weird
         Xt_1 <- c()
-        Xt_1[orders_x != 0] <- gsub(" ", "", paste0("d(", objxvars, ")"), 
+        Xt_1[orders_x != 0] <- gsub(" ", "", paste0("d(", objxvars, ")"),
             fixed = TRUE)[orders_x != 0]
         Xt_1[orders_x == 0] <- gsub(" ", "", objxvars[orders_x == 0], fixed = TRUE)
         if (kw != 0) {
@@ -265,12 +265,12 @@ multipliers.uecm <- function(object, type = "lr", vcov_matrix = NULL) {
     } else { # anything except 0:100 would have been stoped in the earlier check
         interim = type
         Xt_1 <- c()
-        Xt_1[orders_x != 0] <- gsub(" ", "", paste0("L(", objxvars, ", 1)"), 
+        Xt_1[orders_x != 0] <- gsub(" ", "", paste0("L(", objxvars, ", 1)"),
             fixed = TRUE)[orders_x != 0]
         Xt_1[orders_x == 0] <- gsub(" ", "", objxvars[orders_x == 0], fixed = TRUE)
         Xt_1 <- objcoef[objnames_ws %in% Xt_1]
         dXt_interim <- rep(0, length(orders_x))
-        dXt_interim[which(interim < orders_x)] <- objcoef[objnames_ws %in% 
+        dXt_interim[which(interim < orders_x)] <- objcoef[objnames_ws %in%
                     gsub(" ", "", paste0("d(L(", objxvars, ", ", interim, "))"), fixed = TRUE)]
         if (kw != 0) {
             multipliers_coef <- c(